More restructuring addressing comments
