+++ ambry-tools/src/main/java/com.github.ambry/store/DumpDataHelper.java
@@ -13,202 +13,28 @@
  */
 package com.github.ambry.store;
 
-import com.codahale.metrics.MetricRegistry;
 import com.github.ambry.clustermap.ClusterMap;
-import com.github.ambry.clustermap.PartitionId;
 import com.github.ambry.commons.BlobId;
-import com.github.ambry.config.StoreConfig;
-import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.messageformat.BlobData;
 import com.github.ambry.messageformat.BlobProperties;
 import com.github.ambry.messageformat.MessageFormatErrorCodes;
 import com.github.ambry.messageformat.MessageFormatException;
 import com.github.ambry.messageformat.MessageFormatRecord;
 import com.github.ambry.utils.SystemTime;
-import com.github.ambry.utils.Throttler;
 import com.github.ambry.utils.Utils;
 import java.io.DataInputStream;
-import java.io.EOFException;
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
 import java.nio.channels.Channels;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 /**
- * Helper class to assist in dumping data from index or log files
+ * Helper class to assist in dumping a single blob record from a log file
  */
 class DumpDataHelper {
-
-  private final ClusterMap _clusterMap;
-  private Throttler throttler;
-  private static final Logger logger = LoggerFactory.getLogger(DumpDataHelper.class);
-
-  DumpDataHelper(ClusterMap clusterMap, int bytesPerSec) {
-    this._clusterMap = clusterMap;
-    if (bytesPerSec > 0) {
-      this.throttler = new Throttler(bytesPerSec, 100, true, SystemTime.getInstance());
-    } else if (bytesPerSec < 0) {
-      throw new IllegalArgumentException("BytesPerSec " + bytesPerSec + " cannot be negative ");
-    }
-  }
-
-  /**
-   * Dumps all blobs in an index file
-   * @param indexFileToDump the index file that needs to be parsed
-   * @param blobList List of blobIds to be filtered for
-   * @param blobIdToMessageMap {@link Map} of BlobId to {@link IndexRecord} to hold the information
-   *                                          about blobs in the index after parsing
-   * @return the total number of keys/records processed
-   * @throws Exception
-   */
-  long dumpBlobsFromIndex(File indexFileToDump, ArrayList<String> blobList, Map<String, IndexRecord> blobIdToMessageMap)
-      throws Exception {
-    StoreKeyFactory storeKeyFactory = Utils.getObj("com.github.ambry.commons.BlobIdFactory", _clusterMap);
-    StoreConfig config = new StoreConfig(new VerifiableProperties(new Properties()));
-    StoreMetrics metrics = new StoreMetrics(indexFileToDump.getParent(), new MetricRegistry());
-    IndexSegment segment = new IndexSegment(indexFileToDump, false, storeKeyFactory, config, metrics,
-        new Journal(indexFileToDump.getParent(), 0, 0));
-    List<MessageInfo> entries = new ArrayList<>();
-    segment.getEntriesSince(null, new FindEntriesCondition(Long.MAX_VALUE), entries, new AtomicLong(0));
-    long numberOfKeysProcessed = 0;
-    for (MessageInfo entry : entries) {
-      StoreKey key = entry.getStoreKey();
-      IndexValue value = segment.find(key);
-      boolean isDeleted = value.isFlagSet(IndexValue.Flags.Delete_Index);
-      numberOfKeysProcessed++;
-      if (blobList == null || blobList.size() == 0 || blobList.contains(key.toString())) {
-        IndexRecord indexRecord = new IndexRecord(value.toString(), isDeleted, isExpired(value.getExpiresAtMs()));
-        if (blobIdToMessageMap.containsKey(key.getID())) {
-          logger.error(
-              "Duplicate record found for same blob " + key.getID() + ". Prev record " + blobIdToMessageMap.get(
-                  key.getID()) + ", New record " + indexRecord);
-        }
-        blobIdToMessageMap.put(key.getID(), indexRecord);
-      }
-    }
-    return numberOfKeysProcessed;
-  }
-
-  /**
-   * Dumps all blobs in a given log file
-   * @param file the log file that needs to be parsed for
-   * @param startOffset the starting offset from which records needs to be dumped from. Can be {@code null}
-   * @param endOffset the end offset until which records need to be dumped to. Can be {@code null}
-   * @param blobs List of blobIds to be filtered for
-   * @param filter {@code true} if filtering has to be done, {@code false} otherwise
-   * @param blobIdToLogRecord {@link HashMap} of blobId to {@link LogBlobRecord} to hold the information about blobs
-   *                                         in the log after parsing
-   * @param logRecord {@code true} if log record has to be logged, {@code false} otherwise
-   * @throws IOException
-   */
-  void dumpLog(File file, long startOffset, long endOffset, ArrayList<String> blobs, boolean filter,
-      Map<String, LogBlobRecord> blobIdToLogRecord, boolean logRecord) throws IOException {
-    logger.info("Dumping log file " + file.getAbsolutePath());
-    long currentOffset = 0;
-    RandomAccessFile randomAccessFile = new RandomAccessFile(file, "r");
-    long fileSize = file.length();
-    boolean lastBlobFailed = false;
-    if (startOffset != -1) {
-      currentOffset = startOffset;
-    }
-    if (endOffset == -1) {
-      endOffset = fileSize;
-    }
-    logger.info("Starting dumping from offset " + currentOffset);
-    while (currentOffset < endOffset) {
-      try {
-        LogBlobRecordInfo logBlobRecordInfo = readSingleRecordFromLog(randomAccessFile, currentOffset);
-        if (lastBlobFailed) {
-          logger.info("Successful record found at " + currentOffset + " after some failures ");
-        }
-        lastBlobFailed = false;
-        if (!logBlobRecordInfo.isDeleted) {
-          if (filter) {
-            if (blobs.contains(logBlobRecordInfo.blobId.getID())) {
-              if (logRecord) {
-                logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
-                    + logBlobRecordInfo.blobProperty + "\n" + logBlobRecordInfo.userMetadata + "\n"
-                    + logBlobRecordInfo.blobDataOutput);
-              }
-              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
-                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
-            }
-          } else {
-            if (logRecord) {
-              logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
-                  + logBlobRecordInfo.blobProperty + "\n" + logBlobRecordInfo.userMetadata + "\n"
-                  + logBlobRecordInfo.blobDataOutput + " end offset " + (currentOffset
-                  + logBlobRecordInfo.totalRecordSize));
-              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
-                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
-            }
-          }
-        } else {
-          if (filter) {
-            if (blobs.contains(logBlobRecordInfo.blobId.getID())) {
-              if (logRecord) {
-                logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
-                    + logBlobRecordInfo.deleteMsg);
-              }
-              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
-                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
-            }
-          } else {
-            if (logRecord) {
-              logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
-                  + logBlobRecordInfo.deleteMsg + " end offset " + (currentOffset + logBlobRecordInfo.totalRecordSize));
-            }
-            updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
-                !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
-          }
-        }
-        currentOffset += (logBlobRecordInfo.totalRecordSize);
-      } catch (IllegalArgumentException e) {
-        if (!lastBlobFailed) {
-          logger.error("Illegal arg exception thrown at  " + randomAccessFile.getChannel().position() + ", "
-              + "while reading blob starting at offset " + currentOffset + "with exception: " + e.getStackTrace());
-        }
-        currentOffset++;
-        lastBlobFailed = true;
-      } catch (MessageFormatException e) {
-        if (!lastBlobFailed) {
-          logger.error("MessageFormat exception thrown at  " + randomAccessFile.getChannel().position()
-              + " while reading blob starting at offset " + currentOffset + " with exception: " + e.getStackTrace());
-        }
-        currentOffset++;
-        lastBlobFailed = true;
-      } catch (EOFException e) {
-        e.printStackTrace();
-        logger.error("EOFException thrown at " + randomAccessFile.getChannel().position() + ", Cause :" + e.getCause()
-            + ", Msg :" + e.getMessage() + ", stacktrace " + e.getStackTrace());
-        throw (e);
-      } catch (Exception e) {
-        if (!lastBlobFailed) {
-          e.printStackTrace();
-          logger.error(
-              "Unknown exception thrown with Cause " + e.getCause() + ", Msg :" + e.getMessage() + ", stacktrace "
-                  + e.getStackTrace());
-          logger.info("Trying out next offset " + (currentOffset + 1));
-        }
-        currentOffset++;
-        lastBlobFailed = true;
-      }
-    }
-    logger.info("Dumped until offset " + currentOffset);
-  }
 
   /**
    * Fetches one blob record from the log
@@ -218,8 +44,8 @@ class DumpDataHelper {
    * @throws IOException
    * @throws MessageFormatException
    */
-  LogBlobRecordInfo readSingleRecordFromLog(RandomAccessFile randomAccessFile, long currentOffset)
-      throws IOException, MessageFormatException, InterruptedException {
+  static LogBlobRecordInfo readSingleRecordFromLog(RandomAccessFile randomAccessFile, long currentOffset,
+      ClusterMap clusterMap) throws IOException, MessageFormatException {
     String messageheader = null;
     BlobId blobId = null;
     String blobProperty = null;
@@ -246,7 +72,7 @@ class DumpDataHelper {
       totalRecordSize += header.getMessageSize() + buffer.capacity();
       // read blob id
       InputStream streamlog = Channels.newInputStream(randomAccessFile.getChannel());
-      blobId = new BlobId(new DataInputStream(streamlog), _clusterMap);
+      blobId = new BlobId(new DataInputStream(streamlog), clusterMap);
       totalRecordSize += blobId.sizeInBytes();
       if (header.getBlobPropertiesRecordRelativeOffset()
           != MessageFormatRecord.Message_Header_Invalid_Relative_Offset) {
@@ -267,71 +93,38 @@ class DumpDataHelper {
     } else {
       throw new MessageFormatException("Header version not supported " + version, MessageFormatErrorCodes.IO_Error);
     }
-    if (throttler != null) {
-      throttler.maybeThrottle(totalRecordSize);
-    }
     return new LogBlobRecordInfo(messageheader, blobId, blobProperty, usermetadata, blobDataOutput, deleteMsg,
         isDeleted, isExpired, timeToLiveInSeconds, totalRecordSize);
   }
 
   /**
-   * Dumps replica token file
-   * @param replicaTokenFile the file that needs to be parsed for
-   * @throws Exception
+   * Holds information about a single blob record in the log
    */
-  void dumpReplicaToken(File replicaTokenFile) throws Exception {
-    logger.info("Dumping replica token");
-    DataInputStream stream = new DataInputStream(new FileInputStream(replicaTokenFile));
-    short version = stream.readShort();
-    switch (version) {
-      case 0:
-        int Crc_Size = 8;
-        StoreKeyFactory storeKeyFactory = Utils.getObj("com.github.ambry.commons.BlobIdFactory", _clusterMap);
-        FindTokenFactory findTokenFactory =
-            Utils.getObj("com.github.ambry.store.StoreFindTokenFactory", storeKeyFactory);
-        while (stream.available() > Crc_Size) {
-          // read partition id
-          PartitionId partitionId = _clusterMap.getPartitionIdFromStream(stream);
-          // read remote node host name
-          String hostname = Utils.readIntString(stream);
-          // read remote replica path
-          String replicaPath = Utils.readIntString(stream);
-          // read remote port
-          int port = stream.readInt();
-          // read total bytes read from local store
-          long totalBytesReadFromLocalStore = stream.readLong();
-          // read replica token
-          FindToken token = findTokenFactory.getFindToken(stream);
-          logger.info(
-              "partitionId " + partitionId + " hostname " + hostname + " replicaPath " + replicaPath + " port " + port
-                  + " totalBytesReadFromLocalStore " + totalBytesReadFromLocalStore + " token " + token);
-        }
-        logger.info("crc " + stream.readLong());
-    }
-  }
+  static class LogBlobRecordInfo {
+    final String messageHeader;
+    final BlobId blobId;
+    final String blobProperty;
+    final String userMetadata;
+    final String blobDataOutput;
+    final String deleteMsg;
+    final boolean isDeleted;
+    final boolean isExpired;
+    final long timeToLiveInSeconds;
+    final int totalRecordSize;
 
-  /**
-   * Updates the {@link Map} of blobIds to {@link LogBlobRecord} with the information about the passed in
-   * <code>blobId</code>
-   * @param blobIdToLogRecord {@link HashMap} of blobId to {@link LogBlobRecord} that needs to be updated with the
-   *                                         information about the blob
-   * @param blobId the blobId of the blob that needs to be updated in the {@link Map}
-   * @param offset the offset at which the blob record was parsed from in the log file
-   * @param putRecord {@code true} if the record is a Put record, {@code false} otherwise (incase of a Delete record)
-   * @param isExpired {@code true} if the blob has expired, {@code false} otherwise
-   */
-  private void updateBlobIdToLogRecordMap(Map<String, LogBlobRecord> blobIdToLogRecord, String blobId, Long offset,
-      boolean putRecord, boolean isExpired) {
-    if (blobIdToLogRecord != null) {
-      if (blobIdToLogRecord.containsKey(blobId)) {
-        if (putRecord) {
-          blobIdToLogRecord.get(blobId).addPutRecord(offset, isExpired);
-        } else {
-          blobIdToLogRecord.get(blobId).addDeleteRecord(offset);
-        }
-      } else {
-        blobIdToLogRecord.put(blobId, new LogBlobRecord(offset, putRecord, isExpired));
-      }
+    LogBlobRecordInfo(String messageHeader, BlobId blobId, String blobProperty, String userMetadata,
+        String blobDataOutput, String deleteMsg, boolean isDeleted, boolean isExpired, long timeToLiveInSeconds,
+        int totalRecordSize) {
+      this.messageHeader = messageHeader;
+      this.blobId = blobId;
+      this.blobProperty = blobProperty;
+      this.userMetadata = userMetadata;
+      this.blobDataOutput = blobDataOutput;
+      this.deleteMsg = deleteMsg;
+      this.isDeleted = isDeleted;
+      this.isExpired = isExpired;
+      this.timeToLiveInSeconds = timeToLiveInSeconds;
+      this.totalRecordSize = totalRecordSize;
     }
   }
 
@@ -340,129 +133,7 @@ class DumpDataHelper {
    * @param timeToLive time in milliseconds referring to the time to live for the blob
    * @return {@code true} if blob has expired, {@code false} otherwise
    */
-  boolean isExpired(Long timeToLive) {
+  static boolean isExpired(Long timeToLive) {
     return timeToLive != Utils.Infinite_Time && SystemTime.getInstance().milliseconds() > timeToLive;
-  }
-
-  /**
-   * Returns the {@link ClusterMap}
-   * @return the {@link ClusterMap}
-   */
-  ClusterMap getClusterMap() {
-    return this._clusterMap;
-  }
-
-  /**
-   * Holds the information about a single record in the index
-   */
-  class IndexRecord {
-    String message;
-    boolean isDeleted;
-    boolean isExpired;
-
-    IndexRecord(String msg, boolean isDeleted, boolean isExpired) {
-      this.message = msg;
-      this.isDeleted = isDeleted;
-      this.isExpired = isExpired;
-    }
-
-    String getMessage() {
-      return message;
-    }
-
-    boolean isDeleted() {
-      return isDeleted;
-    }
-
-    boolean isExpired() {
-      return isExpired;
-    }
-
-    public String toString() {
-      return message + ", isDeleted:" + isDeleted + ", isExpired:" + isExpired;
-    }
-  }
-
-  /**
-   * Holds information about a blob in the log
-   */
-  class LogBlobRecord {
-    List<Long> putMessageOffsets = new ArrayList<>();
-    List<Long> deleteMessageOffsets = new ArrayList<>();
-    boolean isConsistent;
-    boolean isDeleted;
-    boolean isExpired;
-    boolean duplicatePuts;
-    boolean duplicateDeletes;
-    boolean putAfterDelete;
-
-    LogBlobRecord(Long offset, boolean putRecord, boolean isExpired) {
-      if (putRecord) {
-        putMessageOffsets.add(offset);
-        this.isExpired = isExpired;
-        isConsistent = true;
-      } else {
-        isConsistent = true;
-        isDeleted = true;
-        deleteMessageOffsets.add(offset);
-      }
-    }
-
-    /**
-     * Adds information about a put record
-     * @param offset the offset at which the put record was found
-     * @param isExpired {@code true} if blob is expired, {@code false} otherwise
-     */
-    void addPutRecord(Long offset, boolean isExpired) {
-      isConsistent = false;
-      putMessageOffsets.add(offset);
-      this.isExpired = isExpired;
-      if (putMessageOffsets.size() > 1) {
-        duplicatePuts = true;
-      } else if (deleteMessageOffsets.size() > 0) {
-        putAfterDelete = true;
-      }
-    }
-
-    /**
-     * Adds information about a delete record
-     * @param offset the offset at which the delete record was found
-     */
-    void addDeleteRecord(Long offset) {
-      isDeleted = true;
-      deleteMessageOffsets.add(offset);
-      if (putMessageOffsets.size() == 1 && deleteMessageOffsets.size() == 1) {
-        isConsistent = true;
-      } else if (deleteMessageOffsets.size() > 1) {
-        isConsistent = false;
-        duplicateDeletes = true;
-      }
-    }
-
-    public String toString() {
-      StringBuilder stringBuilder = new StringBuilder();
-      stringBuilder.append("LogBlobRecord:");
-      stringBuilder.append("[isConsistent=" + isConsistent + "],");
-      stringBuilder.append("[isExpired=" + isExpired + "],");
-      stringBuilder.append("[isDeleted=" + isDeleted + "],");
-      if (putMessageOffsets.size() > 0) {
-        stringBuilder.append("[PutMessageOffsets={");
-        for (long putOffset : putMessageOffsets) {
-          stringBuilder.append(putOffset + ",");
-        }
-        stringBuilder.append("}");
-      }
-      if (deleteMessageOffsets.size() > 0) {
-        stringBuilder.append("[DeleteMessageOffsets={");
-        for (long deleteOffset : deleteMessageOffsets) {
-          stringBuilder.append(deleteOffset + ",");
-        }
-        stringBuilder.append("}");
-      }
-      stringBuilder.append("[DuplicatePuts=" + duplicatePuts + "],");
-      stringBuilder.append("[DuplicateDeletes=" + duplicateDeletes + "],");
-      stringBuilder.append("[PutAfterDeletes=" + putAfterDelete + "]");
-      return stringBuilder.toString();
-    }
   }
 }
