+++ ambry-tools/src/main/java/com.github.ambry/store/DumpLogTool.java
@@ -16,73 +16,128 @@ package com.github.ambry.store;
 import com.github.ambry.clustermap.ClusterMap;
 import com.github.ambry.clustermap.ClusterMapManager;
 import com.github.ambry.config.ClusterMapConfig;
+import com.github.ambry.config.Config;
+import com.github.ambry.config.Default;
 import com.github.ambry.config.VerifiableProperties;
-import com.github.ambry.utils.Utils;
+import com.github.ambry.messageformat.MessageFormatException;
+import com.github.ambry.utils.SystemTime;
+import com.github.ambry.utils.Throttler;
+import java.io.EOFException;
 import java.io.File;
 import java.io.IOException;
+import java.io.RandomAccessFile;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
-import joptsimple.ArgumentAcceptingOptionSpec;
-import joptsimple.OptionParser;
-import joptsimple.OptionSet;
+import org.json.JSONException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 /**
- * DumpDataTool tool to assist in dumping data from log file in Ambry
+ * Tool to assist in dumping data from log file in Ambry
  */
 public class DumpLogTool {
+  private final ClusterMap clusterMap;
+  private Throttler throttler = null;
+  /**
+   * Refers to log file that needs to be dumped
+   */
+  @Config("file.to.read")
+  private final String fileToRead;
 
-  private DumpDataHelper dumpDataHelper;
+  /**
+   * File path referring to the hardware layout
+   */
+  @Config("hardware.layout.file.path")
+  private final String hardwareLayoutFilePath;
+
+  /**
+   * File path referring to the partition layout
+   */
+  @Config("partition.layout.file.path")
+  private final String partitionLayoutFilePath;
+
+  /**
+   * List of blobIds (comma separated values) to filter
+   */
+  @Config("blobId.list")
+  private final String blobIdList;
+
+  /**
+   * The offset in the log to start dumping from
+   */
+  @Config("log.start.offset")
+  @Default("0")
+  private final long logStartOffset;
+
+  /**
+   * The offset in the log until which to dump data
+   */
+  @Config("log.end.offset")
+  @Default("-1")
+  private final long logEndOffset;
+
+  /**
+   * The throttling value in bytes per sec
+   */
+  @Config("bytes.per.sec")
+  @Default("0")
+  private final long bytesPerSec;
+
   private static final Logger logger = LoggerFactory.getLogger(DumpLogTool.class);
 
-  public DumpLogTool(ClusterMap map) {
-    dumpDataHelper = new DumpDataHelper(map);
+  private DumpLogTool(VerifiableProperties verifiableProperties) throws IOException, JSONException {
+    fileToRead = verifiableProperties.getString("file.to.read");
+    hardwareLayoutFilePath = verifiableProperties.getString("hardware.layout.file.path");
+    partitionLayoutFilePath = verifiableProperties.getString("partition.layout.file.path");
+    blobIdList = verifiableProperties.getString("blobId.list", "");
+    logStartOffset = verifiableProperties.getInt("log.start.offset", -1);
+    logEndOffset = verifiableProperties.getInt("log.end.offset", -1);
+    bytesPerSec = verifiableProperties.getLong("bytes.per.sec", 0);
+    if (!new File(hardwareLayoutFilePath).exists() || !new File(partitionLayoutFilePath).exists()) {
+      throw new IllegalArgumentException("Hardware or Partition Layout file does not exist");
+    }
+    clusterMap = new ClusterMapManager(hardwareLayoutFilePath, partitionLayoutFilePath,
+        new ClusterMapConfig(new VerifiableProperties(new Properties())));
+    if (bytesPerSec > 0) {
+      this.throttler = new Throttler(bytesPerSec, 100, true, SystemTime.getInstance());
+    } else if (bytesPerSec < 0) {
+      throw new IllegalArgumentException("BytesPerSec " + bytesPerSec + " cannot be negative ");
+    }
   }
 
   public static void main(String args[]) {
     try {
-
-      OptionParser parser = new OptionParser();
-      ArgumentAcceptingOptionSpec<String> propsFileOpt = parser.accepts("propsFile", "Properties file path")
-          .withRequiredArg()
-          .describedAs("propsFile")
-          .ofType(String.class);
-
-      OptionSet options = parser.parse(args);
-      String propsFilePath = options.valueOf(propsFileOpt);
-      Properties properties = Utils.loadProps(propsFilePath);
-      VerifiableProperties verifiableProperties = new VerifiableProperties(properties);
-      DumpLogConfig config = new DumpLogConfig(verifiableProperties);
-
-      if (!new File(config.hardwareLayoutFilePath).exists() || !new File(config.partitionLayoutFilePath).exists()) {
-        throw new IllegalArgumentException("Hardware or Partition Layout file does not exist");
-      }
-
-      ClusterMap map = new ClusterMapManager(config.hardwareLayoutFilePath, config.partitionLayoutFilePath,
-          new ClusterMapConfig(new VerifiableProperties(new Properties())));
-      DumpLogTool dumpLogTool = new DumpLogTool(map);
-
-      boolean filter = !config.blobIdList.equals("");
-      ArrayList<String> blobs = new ArrayList<String>();
-      String[] blobArray;
-      if (filter) {
-        blobArray = config.blobIdList.split(",");
-        blobs.addAll(Arrays.asList(blobArray));
-        logger.info("Blobs to look out for :: " + blobs);
-      }
-
-      if (config.fileToRead != null) {
-        logger.info("File to read " + config.fileToRead);
-      }
-      dumpLogTool.dumpLog(new File(config.fileToRead), config.logStartOffset, config.logEndOffset, blobs, filter);
+      VerifiableProperties verifiableProperties = StoreToolsUtil.getVerifiableProperties(args);
+      DumpLogTool dumpLogTool = new DumpLogTool(verifiableProperties);
+      dumpLogTool.doOperation();
     } catch (Exception e) {
       logger.error("Closed with exception ", e);
     }
+  }
+
+  /**
+   * Executes the operation with the help of properties passed
+   * @throws IOException
+   */
+  private void doOperation() throws IOException {
+    ArrayList<String> blobs = null;
+    String[] blobArray;
+    if (!blobIdList.equals("")) {
+      blobArray = blobIdList.split(",");
+      blobs = new ArrayList<>();
+      blobs.addAll(Arrays.asList(blobArray));
+      logger.info("Blobs to look out for :: " + blobs);
+    }
+
+    if (fileToRead != null) {
+      logger.info("File to read " + fileToRead);
+    }
+    dumpLog(new File(fileToRead), logStartOffset, logEndOffset, blobs);
   }
 
   /**
@@ -90,22 +145,239 @@ public class DumpLogTool {
    * @param logFile the log file that needs to be parsed for
    * @param startOffset the starting offset from which records needs to be dumped from. Can be {@code null}
    * @param endOffset the end offset until which records need to be dumped to. Can be {@code null}
-   * @param blobs List of blobIds to be filtered for
-   * @param filter {@code true} if filtering has to be done, {@code false} otherwise
+   * @param blobs List of blobIds to be filtered for. {@code null} if no filtering required
    * @throws IOException
    */
-  public void dumpLog(File logFile, long startOffset, long endOffset, ArrayList<String> blobs, boolean filter)
-      throws IOException {
-    Map<String, DumpDataHelper.LogBlobRecord> blobIdToLogRecord = new HashMap<>();
-    dumpDataHelper.dumpLog(logFile, startOffset, endOffset, blobs, filter, blobIdToLogRecord, true);
+  private void dumpLog(File logFile, long startOffset, long endOffset, ArrayList<String> blobs) throws IOException {
+    Map<String, LogBlobStatus> blobIdToLogRecord = new HashMap<>();
+    dumpLog(logFile, startOffset, endOffset, blobs, blobIdToLogRecord, true);
     long totalInConsistentBlobs = 0;
     for (String blobId : blobIdToLogRecord.keySet()) {
-      DumpDataHelper.LogBlobRecord logBlobRecord = blobIdToLogRecord.get(blobId);
-      if (!logBlobRecord.isConsistent) {
+      LogBlobStatus logBlobStatus = blobIdToLogRecord.get(blobId);
+      if (!logBlobStatus.isConsistent) {
         totalInConsistentBlobs++;
-        logger.error("Inconsistent blob " + blobId + " " + logBlobRecord);
+        logger.error("Inconsistent blob " + blobId + " " + logBlobStatus);
       }
     }
     logger.info("Total inconsistent blob count " + totalInConsistentBlobs);
+  }
+
+  /**
+   * Dumps all blobs in a given log file
+   * @param file the log file that needs to be parsed for
+   * @param startOffset the starting offset from which records needs to be dumped from. Can be {@code null}
+   * @param endOffset the end offset until which records need to be dumped to. Can be {@code null}
+   * @param blobs List of blobIds to be filtered for. {@code null} if no filtering required
+   * @param blobIdToLogRecord {@link HashMap} of blobId to {@link LogBlobStatus} to hold the information about blobs
+   *                                         in the log after parsing
+   * @param logRecord {@code true} if log record has to be logged, {@code false} otherwise
+   * @throws IOException
+   */
+  private void dumpLog(File file, long startOffset, long endOffset, ArrayList<String> blobs,
+      Map<String, LogBlobStatus> blobIdToLogRecord, boolean logRecord) throws IOException {
+    logger.info("Dumping log file " + file.getAbsolutePath());
+    long currentOffset = 0;
+    RandomAccessFile randomAccessFile = new RandomAccessFile(file, "r");
+    long fileSize = file.length();
+    boolean lastBlobFailed = false;
+    if (startOffset != -1) {
+      currentOffset = startOffset;
+    }
+    if (endOffset == -1) {
+      endOffset = fileSize;
+    }
+    logger.info("Starting dumping from offset " + currentOffset);
+    while (currentOffset < endOffset) {
+      try {
+        DumpDataHelper.LogBlobRecordInfo logBlobRecordInfo =
+            DumpDataHelper.readSingleRecordFromLog(randomAccessFile, currentOffset, clusterMap);
+        if (throttler != null) {
+          throttler.maybeThrottle(logBlobRecordInfo.totalRecordSize);
+        }
+        if (lastBlobFailed) {
+          logger.info("Successful record found at " + currentOffset + " after some failures ");
+        }
+        lastBlobFailed = false;
+        if (!logBlobRecordInfo.isDeleted) {
+          if (blobs != null) {
+            if (blobs.contains(logBlobRecordInfo.blobId.getID())) {
+              if (logRecord) {
+                logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
+                    + logBlobRecordInfo.blobProperty + "\n" + logBlobRecordInfo.userMetadata + "\n"
+                    + logBlobRecordInfo.blobDataOutput);
+              }
+              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
+                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
+            }
+          } else {
+            if (logRecord) {
+              logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
+                  + logBlobRecordInfo.blobProperty + "\n" + logBlobRecordInfo.userMetadata + "\n"
+                  + logBlobRecordInfo.blobDataOutput + " end offset " + (currentOffset
+                  + logBlobRecordInfo.totalRecordSize));
+              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
+                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
+            }
+          }
+        } else {
+          if (blobs != null) {
+            if (blobs.contains(logBlobRecordInfo.blobId.getID())) {
+              if (logRecord) {
+                logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
+                    + logBlobRecordInfo.deleteMsg);
+              }
+              updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
+                  !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
+            }
+          } else {
+            if (logRecord) {
+              logger.info(logBlobRecordInfo.messageHeader + "\n " + logBlobRecordInfo.blobId + "\n"
+                  + logBlobRecordInfo.deleteMsg + " end offset " + (currentOffset + logBlobRecordInfo.totalRecordSize));
+            }
+            updateBlobIdToLogRecordMap(blobIdToLogRecord, logBlobRecordInfo.blobId.getID(), currentOffset,
+                !logBlobRecordInfo.isDeleted, logBlobRecordInfo.isExpired);
+          }
+        }
+        currentOffset += (logBlobRecordInfo.totalRecordSize);
+      } catch (IllegalArgumentException e) {
+        if (!lastBlobFailed) {
+          logger.error("Illegal arg exception thrown at  " + randomAccessFile.getChannel().position() + ", "
+              + "while reading blob starting at offset " + currentOffset + "with exception: ", e);
+        }
+        currentOffset++;
+        lastBlobFailed = true;
+      } catch (MessageFormatException e) {
+        if (!lastBlobFailed) {
+          logger.error("MessageFormat exception thrown at  " + randomAccessFile.getChannel().position()
+              + " while reading blob starting at offset " + currentOffset + " with exception: ", e);
+        }
+        currentOffset++;
+        lastBlobFailed = true;
+      } catch (EOFException e) {
+        e.printStackTrace();
+        logger.error("EOFException thrown at " + randomAccessFile.getChannel().position() + ", Cause :" + e.getCause()
+            + ", Msg :" + e.getMessage() + ", stacktrace ", e);
+        throw (e);
+      } catch (Exception e) {
+        if (!lastBlobFailed) {
+          e.printStackTrace();
+          logger.error(
+              "Unknown exception thrown with Cause " + e.getCause() + ", Msg :" + e.getMessage() + ", stacktrace ", e);
+          logger.info("Trying out next offset " + (currentOffset + 1));
+        }
+        currentOffset++;
+        lastBlobFailed = true;
+      }
+    }
+    logger.info("Dumped until offset " + currentOffset);
+  }
+
+  /**
+   * Updates the {@link Map} of blobIds to {@link LogBlobStatus} with the information about the passed in
+   * <code>blobId</code>
+   * @param blobIdToLogRecord {@link HashMap} of blobId to {@link LogBlobStatus} that needs to be updated with the
+   *                                         information about the blob
+   * @param blobId the blobId of the blob that needs to be updated in the {@link Map}
+   * @param offset the offset at which the blob record was parsed from in the log file
+   * @param putRecord {@code true} if the record is a Put record, {@code false} otherwise (incase of a Delete record)
+   * @param isExpired {@code true} if the blob has expired, {@code false} otherwise
+   */
+  private void updateBlobIdToLogRecordMap(Map<String, LogBlobStatus> blobIdToLogRecord, String blobId, Long offset,
+      boolean putRecord, boolean isExpired) {
+    if (blobIdToLogRecord != null) {
+      if (blobIdToLogRecord.containsKey(blobId)) {
+        if (putRecord) {
+          blobIdToLogRecord.get(blobId).addPutRecord(offset, isExpired);
+        } else {
+          blobIdToLogRecord.get(blobId).addDeleteRecord(offset);
+        }
+      } else {
+        blobIdToLogRecord.put(blobId, new LogBlobStatus(offset, putRecord, isExpired));
+      }
+    }
+  }
+
+  /**
+   * Holds information about a blob in the log. If a multiple records are found for the same blob, its captured
+   * in the same instance of this class.
+   */
+  private class LogBlobStatus {
+    List<Long> putMessageOffsets = new ArrayList<>();
+    List<Long> deleteMessageOffsets = new ArrayList<>();
+    boolean isConsistent;
+    boolean isDeleted;
+    boolean isExpired;
+    boolean duplicatePuts;
+    boolean duplicateDeletes;
+    boolean putAfterDelete;
+
+    private LogBlobStatus(Long offset, boolean putRecord, boolean isExpired) {
+      if (putRecord) {
+        putMessageOffsets.add(offset);
+        this.isExpired = isExpired;
+        isConsistent = true;
+      } else {
+        isConsistent = true;
+        isDeleted = true;
+        deleteMessageOffsets.add(offset);
+      }
+    }
+
+    /**
+     * Adds information about a put record
+     * @param offset the offset at which the put record was found
+     * @param isExpired {@code true} if blob is expired, {@code false} otherwise
+     */
+    void addPutRecord(Long offset, boolean isExpired) {
+      isConsistent = false;
+      putMessageOffsets.add(offset);
+      this.isExpired = isExpired;
+      if (putMessageOffsets.size() > 1) {
+        duplicatePuts = true;
+      } else if (deleteMessageOffsets.size() > 0) {
+        putAfterDelete = true;
+      }
+    }
+
+    /**
+     * Adds information about a delete record
+     * @param offset the offset at which the delete record was found
+     */
+    void addDeleteRecord(Long offset) {
+      isDeleted = true;
+      deleteMessageOffsets.add(offset);
+      if (putMessageOffsets.size() == 1 && deleteMessageOffsets.size() == 1) {
+        isConsistent = true;
+      } else if (deleteMessageOffsets.size() > 1) {
+        isConsistent = false;
+        duplicateDeletes = true;
+      }
+    }
+
+    public String toString() {
+      StringBuilder stringBuilder = new StringBuilder();
+      stringBuilder.append("LogBlobStatus:");
+      stringBuilder.append("[isConsistent=" + isConsistent + "],");
+      stringBuilder.append("[isExpired=" + isExpired + "],");
+      stringBuilder.append("[isDeleted=" + isDeleted + "],");
+      if (putMessageOffsets.size() > 0) {
+        stringBuilder.append("[PutMessageOffsets={");
+        for (long putOffset : putMessageOffsets) {
+          stringBuilder.append(putOffset + ",");
+        }
+        stringBuilder.append("}");
+      }
+      if (deleteMessageOffsets.size() > 0) {
+        stringBuilder.append("[DeleteMessageOffsets={");
+        for (long deleteOffset : deleteMessageOffsets) {
+          stringBuilder.append(deleteOffset + ",");
+        }
+        stringBuilder.append("}");
+      }
+      stringBuilder.append("[DuplicatePuts=" + duplicatePuts + "],");
+      stringBuilder.append("[DuplicateDeletes=" + duplicateDeletes + "],");
+      stringBuilder.append("[PutAfterDeletes=" + putAfterDelete + "]");
+      return stringBuilder.toString();
+    }
   }
 }
