+++ ambry-tools/src/main/java/com.github.ambry/store/DumpReplicaTokenTool.java
@@ -15,14 +15,17 @@ package com.github.ambry.store;
 
 import com.github.ambry.clustermap.ClusterMap;
 import com.github.ambry.clustermap.ClusterMapManager;
+import com.github.ambry.clustermap.PartitionId;
 import com.github.ambry.config.ClusterMapConfig;
+import com.github.ambry.config.Config;
 import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.utils.Utils;
+import java.io.DataInputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.util.Properties;
-import joptsimple.ArgumentAcceptingOptionSpec;
-import joptsimple.OptionParser;
-import joptsimple.OptionSet;
+import org.json.JSONException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,42 +35,82 @@ import org.slf4j.LoggerFactory;
  */
 public class DumpReplicaTokenTool {
 
-  private DumpDataHelper dumpDataHelper;
+  private final ClusterMap clusterMap;
+  /**
+   * Refers to replicatoken file that needs to be dumped
+   */
+  @Config("file.to.read")
+  public final String fileToRead;
+
+  /**
+   * File path referring to the hardware layout
+   */
+  @Config("hardware.layout.file.path")
+  public final String hardwareLayoutFilePath;
+
+  /**
+   * File path referring to the partition layout
+   */
+  @Config("partition.layout.file.path")
+  public final String partitionLayoutFilePath;
+
   private static final Logger logger = LoggerFactory.getLogger(DumpDataTool.class);
 
-  public DumpReplicaTokenTool(ClusterMap map) {
-    dumpDataHelper = new DumpDataHelper(map);
+  public DumpReplicaTokenTool(VerifiableProperties verifiableProperties) throws IOException, JSONException {
+
+    fileToRead = verifiableProperties.getString("file.to.read");
+    hardwareLayoutFilePath = verifiableProperties.getString("hardware.layout.file.path");
+    partitionLayoutFilePath = verifiableProperties.getString("partition.layout.file.path");
+
+    if (!new File(hardwareLayoutFilePath).exists() || !new File(partitionLayoutFilePath).exists()) {
+      throw new IllegalArgumentException("Hardware or Partition Layout file does not exist");
+    }
+    clusterMap = new ClusterMapManager(hardwareLayoutFilePath, partitionLayoutFilePath,
+        new ClusterMapConfig(new VerifiableProperties(new Properties())));
   }
 
   public static void main(String args[]) {
     try {
-
-      OptionParser parser = new OptionParser();
-      ArgumentAcceptingOptionSpec<String> propsFileOpt = parser.accepts("propsFile", "Properties file path")
-          .withRequiredArg()
-          .describedAs("propsFile")
-          .ofType(String.class);
-
-      OptionSet options = parser.parse(args);
-      String propsFilePath = options.valueOf(propsFileOpt);
-      Properties properties = Utils.loadProps(propsFilePath);
-      VerifiableProperties verifiableProperties = new VerifiableProperties(properties);
-      DumpReplicaTokenConfig config = new DumpReplicaTokenConfig(verifiableProperties);
-
-      if (!new File(config.hardwareLayoutFilePath).exists() || !new File(config.partitionLayoutFilePath).exists()) {
-        throw new IllegalArgumentException("Hardware or Partition Layout file does not exist");
-      }
-      ClusterMap map = new ClusterMapManager(config.hardwareLayoutFilePath, config.partitionLayoutFilePath,
-          new ClusterMapConfig(new VerifiableProperties(new Properties())));
-      DumpReplicaTokenTool dumpReplicaTokenTool = new DumpReplicaTokenTool(map);
-
-      if (config.fileToRead != null) {
-        logger.info("File to read " + config.fileToRead);
-      }
-
-      dumpReplicaTokenTool.dumpDataHelper.dumpReplicaToken(new File(config.fileToRead));
+      VerifiableProperties verifiableProperties = StoreToolsUtil.getVerifiableProperties(args);
+      DumpReplicaTokenTool dumpReplicaTokenTool = new DumpReplicaTokenTool(verifiableProperties);
+      dumpReplicaTokenTool.dumpReplicaToken();
     } catch (Exception e) {
       logger.error("Closed with exception ", e);
+    }
+  }
+
+  /**
+   * Dumps replica token file
+   * @throws Exception
+   */
+  private void dumpReplicaToken() throws Exception {
+    logger.info("Dumping replica token file " + fileToRead);
+    DataInputStream stream = new DataInputStream(new FileInputStream(fileToRead));
+    short version = stream.readShort();
+    switch (version) {
+      case 0:
+        int Crc_Size = 8;
+        StoreKeyFactory storeKeyFactory = Utils.getObj("com.github.ambry.commons.BlobIdFactory", clusterMap);
+        FindTokenFactory findTokenFactory =
+            Utils.getObj("com.github.ambry.store.StoreFindTokenFactory", storeKeyFactory);
+        while (stream.available() > Crc_Size) {
+          // read partition id
+          PartitionId partitionId = clusterMap.getPartitionIdFromStream(stream);
+          // read remote node host name
+          String hostname = Utils.readIntString(stream);
+          // read remote replica path
+          String replicaPath = Utils.readIntString(stream);
+          // read remote port
+          int port = stream.readInt();
+          // read total bytes read from local store
+          long totalBytesReadFromLocalStore = stream.readLong();
+          // read replica token
+          FindToken token = findTokenFactory.getFindToken(stream);
+          logger.info(
+              "partitionId " + partitionId + " hostname " + hostname + " replicaPath " + replicaPath + " port " + port
+                  + " totalBytesReadFromLocalStore " + totalBytesReadFromLocalStore + " token " + token);
+        }
+        logger.info("crc " + stream.readLong());
     }
   }
 }
